+++ Modules/Ruleset.Compatibility/Public/Compare-WinModule.ps1
@@ -5,7 +5,7 @@ MIT License
 This file is part of "Windows Firewall Ruleset" project
 Homepage: https://github.com/metablaster/WindowsFirewallRuleset
 
-Copyright (C) 2020 metablaster zebal@protonmail.ch
+Copyright (C) 2018, 2019 Microsoft Corporation. All rights reserved
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -33,6 +33,21 @@ Compare the set of modules for this version of PowerShell against those availabl
 .DESCRIPTION
 Compare the set of modules for this version of PowerShell against those available in the compatibility session.
 
+.PARAMETER Name
+Specifies the names or name patterns of for the modules to compare.
+Wildcard characters are permitted.
+
+.PARAMETER ComputerName
+If you don't want to use the default compatibility session, use this parameter to specify the name
+of the computer on which to create the compatibility session.
+
+.PARAMETER ConfigurationName
+Specifies the configuration to connect to when creating the compatibility session
+(Defaults to 'Microsoft.PowerShell')
+
+.PARAMETER Credential
+If needed, use this parameter to specify credentials for the compatibility session
+
 .EXAMPLE
 PS> Compare-WinModule
 
@@ -52,36 +67,30 @@ None. You cannot pipe objects to Compare-WinModule
 
 .NOTES
 None.
-TODO: Update Copyright and start implementing module function
-TODO: Update HelpURI
+
+.LINK
+https://github.com/PowerShell/WindowsCompatibility
 #>
 function Compare-WinModule
 {
-	[CmdletBinding()]
+	[CmdletBinding(
+		HelpURI = "https://github.com/metablaster/WindowsFirewallRuleset/blob/master/Modules/Ruleset.Compatibility/Help/en-US/Compare-WinModule.md")]
 	[OutputType([PSObject])]
 	Param
 	(
-		# Specifies the names or name patterns of for the modules to compare.
-		# Wildcard characters are permitted.
 		[Parameter(Position = 0)]
 		[String[]]
 		$Name = "*",
 
-		# If you don't want to use the default compatibility session, use
-		# this parameter to specify the name of the computer on which to create
-		# the compatibility session.
 		[Parameter()]
 		[String]
 		[Alias("cn")]
 		$ComputerName,
 
-		# Specifies the configuration to connect to when creating the compatibility session
-		# (Defaults to 'Microsoft.PowerShell')
 		[Parameter()]
 		[String]
 		$ConfigurationName,
 
-		# If needed, use this parameter to specify credentials for the compatibility session
 		[Parameter()]
 		[PSCredential]
 		$Credential