+++ Modules/Ruleset.Compatibility/Public/Invoke-WinCommand.ps1
@@ -5,7 +5,7 @@ MIT License
 This file is part of "Windows Firewall Ruleset" project
 Homepage: https://github.com/metablaster/WindowsFirewallRuleset
 
-Copyright (C) 2020 metablaster zebal@protonmail.ch
+Copyright (C) 2018, 2019 Microsoft Corporation. All rights reserved
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -38,6 +38,23 @@ By default, when executing, the current compatibility session is used,
 or, in the case where there is no existing session, a new default session will be created.
 This behavior can be overridden using the additional parameters on the command.
 
+.PARAMETER ScriptBlock
+The scriptblock to invoke in the compatibility session
+
+.PARAMETER ComputerName
+If you don't want to use the default compatibility session, use this parameter to specify the name
+of the computer on which to create the compatibility session.
+
+.PARAMETER ConfigurationName
+Specifies the configuration to connect to when creating the compatibility session
+(Defaults to 'Microsoft.PowerShell')
+
+.PARAMETER Credential
+The credential to use when connecting to the compatibility session.
+
+.PARAMETER ArgumentList
+Arguments to pass to the scriptblock
+
 .EXAMPLE
 PS> Invoke-WinCommand {param ($name) "Hello $name, how are you?"; $PSVersionTable.PSVersion} Jeffrey
 
@@ -63,40 +80,34 @@ System.Management.Automation.PSObject
 
 .NOTES
 None.
-TODO: Update Copyright and start implementing module function
-TODO: Update HelpURI
+
+.LINK
+https://github.com/PowerShell/WindowsCompatibility
 #>
 function Invoke-WinCommand
 {
-	[CmdletBinding()]
+	[CmdletBinding(
+		HelpURI = "https://github.com/metablaster/WindowsFirewallRuleset/blob/master/Modules/Ruleset.Compatibility/Help/en-US/Invoke-WinCommand.md")]
 	[OutputType([PSObject])]
 	Param
 	(
-		# The scriptblock to invoke in the compatibility session
 		[Parameter(Mandatory, Position = 0)]
 		[ScriptBlock]
 		$ScriptBlock,
 
-		# If you don't want to use the default compatibility session, use
-		# this parameter to specify the name of the computer on which to create
-		# the compatibility session.
 		[Parameter()]
 		[String]
 		[Alias("cn")]
 		$ComputerName,
 
-		# Specifies the configuration to connect to when creating the compatibility session
-		# (Defaults to 'Microsoft.PowerShell')
 		[Parameter()]
 		[String]
 		$ConfigurationName,
 
-		# The credential to use when connecting to the compatibility session.
 		[Parameter()]
 		[PSCredential]
 		$Credential,
 
-		# Arguments to pass to the scriptblock
 		[Parameter(ValueFromRemainingArguments)]
 		[object[]]
 		$ArgumentList