+++ Modules/Ruleset.Compatibility/Ruleset.Compatibility.psm1
@@ -1,8 +1,30 @@
-###########################################################################################
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License.
 
-# TODO: Update Copyright and start writing code
+<#
+MIT License
+
+This file is part of "Windows Firewall Ruleset" project
+Homepage: https://github.com/metablaster/WindowsFirewallRuleset
+
+Copyright (C) 2018, 2019 Microsoft Corporation. All rights reserved
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+#>
 
 # Initialization
 using namespace System.Management.Automation.
@@ -16,17 +38,7 @@ Set-Variable -Name ThisModule -Scope Script -Option ReadOnly -Force -Value ($MyI
 
 #
 # Script imports
-# TODO: which imports should be private?
 #
-
-$PrivateScripts = @(
-)
-
-foreach ($Script in $PrivateScripts)
-{
-	Write-Debug -Message "[$ThisModule] Importing script: Private\$Script.ps1"
-	. ("{0}\Private\{1}.ps1" -f $PSScriptRoot, $Script)
-}
 
 $PublicScripts = @(
 	"Add-WindowsPSModulePath"
@@ -57,7 +69,7 @@ Set-Variable -Name NeverImportList -Scope Script -Value @(
 	"PowerShellGet",
 	"Microsoft.PowerShell.Archive",
 	"Microsoft.PowerShell.Host",
-	"WindowsCompatibility"
+	"Ruleset.Compatibility"
 )
 
 ###########################################################################################