+++ Modules/Ruleset.Compatibility/Test/PSModulePath.Tests.ps1
@@ -1,16 +1,17 @@
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License.
+
+# Copyright (C) 2018, 2019 Microsoft Corporation. All rights reserved
+# Licensed under the MIT License
 
 $scriptPath = Split-Path $MyInvocation.MyCommand.Path -Parent
 
 Describe "Test Add-WindowsPSModulePath cmdlet" {
 	BeforeAll {
-		Import-Module -Force "$scriptPath\..\bin\WindowsCompatibility.psd1"
+		Import-Module -Force "$scriptPath\..\Ruleset.Compatibility.psd1"
 		$originalPsModulePath = $env:PSModulePath
 	}
 
 	AfterAll {
-		Remove-Module -Force WindowsCompatibility
+		Remove-Module -Force Ruleset.Compatibility
 		$env:PSModulePath = $originalPsModulePath
 	}